greentea/tgvisd: Implement chat lock by chat_id
Note that when we're working with database, we may have multiple
threads doing work to the same chat_id, thus it can cause race
condition problem. Currently we don't have a good workaround for
the MySQL side, so let's create a map lock mutex on the app, this
should have better latency as well, as locking on MySQL will
require network communication, which is very likely slower than
SMP synchronization.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>